git-svn-id: http://southwind:2481/svn/cpp_lib@11 287b3242-7fab-264f-8401-8509467ab285

+++ math/Sieve.hpp
@@ -52,7 +52,7 @@ private:
 				return offset;
 		}
 
-		return bitSieve.size();
+		return static_cast<unsigned int>(bitSieve.size());
 	}
 	
 	void mark(unsigned int offset, unsigned int step)
